+++ packages/network-controller/tests/provider-api-tests/block-hash-in-response.ts
@@ -365,25 +365,26 @@ export function testsForRpcMethodsThatCheckForBlockHashInResponse(
         });
       });
 
-      testsForRpcFailoverBehavior({
-        providerType,
-        requestToCall: {
-          method,
-          params: [],
-        },
-        getRequestToMock: () => ({
-          method,
-          params: [],
-        }),
-        failure: {
-          httpStatus,
-        },
-        isRetriableFailure: false,
-        getExpectedError: () =>
-          expect.objectContaining({
-            message: errorMessage,
-          }),
-      });
+      // TODO: Add tests for failover behavior when the RPC endpoint returns a 405 or 429 response without opening a circuit breaker
+      // testsForRpcFailoverBehavior({
+      //   providerType,
+      //   requestToCall: {
+      //     method,
+      //     params: [],
+      //   },
+      //   getRequestToMock: () => ({
+      //     method,
+      //     params: [],
+      //   }),
+      //   failure: {
+      //     httpStatus,
+      //   },
+      //   isRetriableFailure: false,
+      //   getExpectedError: () =>
+      //     expect.objectContaining({
+      //       message: errorMessage,
+      //     }),
+      // });
     },
   );